requirements.txt
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.9.3
Automat==0.7.0
Babel==2.7.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
bleach==3.1.0
blinker==1.4
blosc==1.8.1
boto==2.49.0
boto3==1.9.213
botocore==1.12.213
Bottleneck==1.2.1
bower.py==1.0
certifi==2019.6.16
cffi==1.12.3
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
chart==0.2.3
cssselect==1.0.3
cssselect2==0.2.2
csv-to-sqlite==2.0.0
cycler==0.10.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.4
django-admin-tools==0.8.1
django-admin-tools-stats==0.9.0
django-autoslug-iplweb==1.9.4
django-bower==5.2.0
django-braces==1.13.0
django-chartit==0.2.9
django-chartjs==1.5.0
django-chatter==1.0.7
django-controlcenter==0.2.9
django-crispy-forms==1.7.2
django-dash==0.5.6
django-filter==2.2.0
django-filters==0.2.1
django-fluent-dashboard==1.0.1
django-grappelli==2.13.1
django-heroku==0.3.1
django-jet==1.0.8
django-jsonfield==1.3.1
django-nine==0.2.2
django-nvd3==0.9.7
django-oauth-toolkit==1.2.0
django-pkgconf==0.3.0
django-qsstats-magic==1.1.0
django-report-builder==6.3.2
django-report-tools==0.2.2
django-reporting==0.21
django-reset-migrations==0.3.1
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-rest-multiple-models==2.1.3
django-rest-serializer-field-permissions==2.0.0
django-restful-admin==1.0.3
django-tinymce==2.8.0
djangorestframework==3.10.2
djcacheutils==3.0.0
dnspython==1.16.0
docker==4.0.2
docker-pycreds==0.4.0
docutils==0.14
easy-thumbnails==2.6
ecdsa==0.13.2
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.15
fastparquet==0.3.2
feedparser==5.2.1
fsspec==0.4.0
fusepy==3.0.1
future==0.17.1
gast==0.2.2
gcsfs==0.3.0
Genshi==0.7.3
geocoder==1.38.1
geoip2==2.9.0
get==2019.4.13
gevent==1.4.0
gitdb2==2.0.5
github3.py==1.3.0
GitPython==2.1.11
gmplot==1.2.0
google-api-python-client==1.4.1
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
google-pasta==0.1.7
googleapis-common-protos==1.6.0
googlemaps==3.0.2
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.13.1
hyperframe==5.2.0
hyperlink==19.0.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.1.1
numpy==1.17.1
oauth2client==4.1.3
oauthlib==3.1.0
plotly==3.10.0
ptyprocess==0.6.0
public==2019.4.13
py==1.8.0
pycares==3.0.0
pycparser==2.19
pyparsing==2.4.2
pyPdf==1.13
PyPDF2==1.26.0
python-dateutil==2.8.0
python-gflags==3.1.2
python-memcached==1.58
python-mimeparse==1.6.0
python-nvd3==0.14.2
python-pptx==0.6.5
python-slugify==1.1.4
python3-openid==3.1.0
reportlab==3.5.23
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.2.0
rest-condition==1.0.3
service-identity==18.1.0
simplejson==3.16.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
SQLAlchemy==1.3.7
sqlparse==0.3.0
Unidecode==1.1.1
WeasyPrint==48
websocket-client==0.56.0
wget==3.2
whitenoise==4.1.3
xhtml2pdf==0.2.3
